Move Results to the only place it is used.
