+++ container-search/src/main/java/com/yahoo/search/federation/FederationSearcher.java
@@ -9,8 +9,7 @@ import com.yahoo.component.chain.Chain;
 import com.yahoo.component.chain.dependencies.After;
 import com.yahoo.component.chain.dependencies.Provides;
 import com.yahoo.component.provider.ComponentRegistry;
-import com.yahoo.errorhandling.Results;
-import com.yahoo.errorhandling.Results.Builder;
+import com.yahoo.search.federation.Results.Builder;
 import com.yahoo.processing.IllegalInputException;
 import com.yahoo.processing.request.CompoundName;
 import com.yahoo.search.Query;
@@ -88,7 +87,7 @@ public class FederationSearcher extends ForkingSearcher {
 
     @Inject
     public FederationSearcher(FederationConfig config, SchemaInfo schemaInfo,
-                              ComponentRegistry<TargetSelector> targetSelectors) {
+                              ComponentRegistry<TargetSelector<?>> targetSelectors) {
         this(createResolver(config),
              createVirtualSourceResolver(config),
              resolveSelector(config.targetSelector(), targetSelectors),
@@ -103,7 +102,7 @@ public class FederationSearcher extends ForkingSearcher {
 
     private FederationSearcher(SearchChainResolver searchChainResolver,
                                VirtualSourceResolver virtualSourceResolver,
-                               TargetSelector targetSelector,
+                               TargetSelector<?> targetSelector,
                                Map<String, List<String>> schema2Clusters) {
         this.searchChainResolver = searchChainResolver;
         sourceRefResolver = new SourceRefResolver(searchChainResolver, schema2Clusters);
@@ -115,8 +114,8 @@ public class FederationSearcher extends ForkingSearcher {
         return VirtualSourceResolver.of(config.target().stream().map(FederationConfig.Target::id).collect(Collectors.toUnmodifiableSet()));
     }
 
-    private static TargetSelector resolveSelector(String selectorId,
-                                                  ComponentRegistry<TargetSelector> targetSelectors) {
+    private static TargetSelector<?> resolveSelector(String selectorId,
+                                                  ComponentRegistry<TargetSelector<?>> targetSelectors) {
         if (selectorId.isEmpty()) return null;
         return checkNotNull(targetSelectors.getComponent(selectorId),
                 "Missing target selector with id '" + selectorId + "'");
