+++ container-search/src/main/java/com/yahoo/search/federation/Results.java
@@ -1,5 +1,5 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.errorhandling;
+package com.yahoo.search.federation;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -13,13 +13,9 @@ public class Results<DATA, ERROR> {
     private final List<DATA> data;
     private final List<ERROR> errors;
 
-    public Results(List<DATA> data, List<ERROR> errors) {
+    private Results(List<DATA> data, List<ERROR> errors) {
         this.data = List.copyOf(data);
         this.errors = List.copyOf(errors);
-    }
-
-    public boolean hasErrors() {
-        return !errors.isEmpty();
     }
 
     public List<DATA> data() {
@@ -44,10 +40,6 @@ public class Results<DATA, ERROR> {
 
         public void addError(ERROR e) {
             errors.add(e);
-        }
-
-        public void addAllErrors(Collection<? extends ERROR> e) {
-            errors.addAll(e);
         }
 
         public Results<DATA, ERROR> build() {
