--- vespajlib/src/main/java/com/yahoo/errorhandling/package-info.java
@@ -1,5 +1,0 @@
-// Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-@ExportPackage
-package com.yahoo.errorhandling;
-
-import com.yahoo.osgi.annotation.ExportPackage;
